requirements.txt
pandas==2.3.0
scikit-learn==1.7.0
xgboost==3.0.2
matplotlib==3.10.3
openpyxl==3.1.5
google-cloud-storage==3.1.0